Using a more sane Dataloader loading mechanism
The dataloader/1 helper seems to put the id in the batch:

[batch: {:team, %{id: "5"}}, ids: [%{}]]

This makes the loader behavior a bit odd because the result was always a
map with a key of %{} (because thats what the args were). Instead a
custom load function is used so that the batch is just the resource and
the args are passed in properly (as a map of args for the field):

[batch: :team, args: [%{id: "5"}]]

This makes the fetch/2 method a little more clearer. It now needs to
return a map with keys from the args.